+++ messanger/src/main/java/com/dimka228/messanger/repositories/UserInChatRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 @Repository
-public interface UserInChatRepository extends JpaRepository<UserInChat, Long> {
+public interface UserInChatRepository extends JpaRepository<UserInChat, Integer> {
     Optional<UserInChat> findByUserIdAndChatId(Integer userId, Integer chatId);
     Optional<List<UserInChat>> findAllByChatId(Integer id);
 }
